Completed

+++ app/Http/Controllers/ListingController.php
@@ -50,6 +50,9 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logos', 'public')->store();
         }
 
+        // Tie Listings to User
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         // Session::flash('message', 'Listing Created');
@@ -64,7 +67,14 @@ class ListingController extends Controller
 
       // Update Listing Data
       public function update(Request $request, Listing $listing) {
+
         // dd($request->file('logo')->store());
+
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -89,9 +99,23 @@ class ListingController extends Controller
 
     // Delete Listing
     public function destroy(Listing $listing) {
+        
+        // Make sure logged in user is owner
+        if($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully!');
     }
+
+
+    // Manage Listings
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
+    }
+
+
 }
